requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.0.0
faiss-cpu
python-dotenv
tiktoken
